requirements.txt
pycrypto
six
scrypt
